- Relative path bug fixed - Changing no data values in temp HPMF and ISI layers to np.mean instead of hardcoded values (hardcoded values are still used if the tile is full of nans) - The script now accepts any size input DEM as long as both sides are at least 512 pixels - Additional depth map creation, depths are adopt from HPMF layer

+++ U-Net_PyTorch/inference.py
@@ -11,34 +11,44 @@ import rasterio
 from utils import create_hpmf_layer, create_isi_layer, create_feature_layer
 from model import DitchNet
 
+from osgeo import gdal
+gdal.UseExceptions()
+
 
 class DitchNetPredictor:
     def __init__(self, model, input_dem_dir, output_dir, threshold,
-                 output_prob_map=True, output_class_map=True, device=None):
+                 output_prob_map=True, output_class_map=True, output_depth_map=True,
+                 device=None):
 
-        if not output_prob_map and not output_class_map:
-            raise ValueError('At least one of "output_prob_map" or "output_class_map" must be True.')
+        if not output_prob_map and not output_class_map and not output_depth_map:
+            raise ValueError('At least one of "output_prob_map", "output_class_map" or "output_depth_map" must be True.')
 
-        self.input_dem_dir = Path(input_dem_dir)
-        self.output_dir = Path(output_dir)
+        self.input_dem_dir = Path(input_dem_dir).resolve()
+        self.output_dir = Path(output_dir).resolve()
         self.output_dir.mkdir(parents=True, exist_ok=True)
 
         self.threshold = threshold
 
         if device is None:
             device = "cuda" if torch.cuda.is_available() else "cpu"
+
         self.device = torch.device(device)
-        print(f"Using device: {self.device}")
+        print(f"\nUsing device: {self.device}")
 
         self.model = model.to(self.device)
         self.model.eval()
 
-        self.create_prob_map, self.create_class_map = output_prob_map, output_class_map
-        self.output_probability_dir, self.output_classified_dir = None, None
+        self.create_prob_map = output_prob_map
+        self.create_class_map = output_class_map
+        self.create_depth_map = output_depth_map
+
+        self.output_probability_dir, self.output_classified_dir, self.output_depth_dir = None, None, None
         self._set_output_directories()
 
         self.temp_dir, self.hpmf_temp, self.isi_temp = None, None, None
         self._set_temporary_directories()
+
+        self.invalid_inputs = []
 
     def _set_output_directories(self):
         if self.create_prob_map:
@@ -49,6 +59,10 @@ class DitchNetPredictor:
             self.output_classified_dir = self.output_dir / "classified_maps"
             self.output_classified_dir.mkdir(parents=True, exist_ok=True)
 
+        if self.create_depth_map:
+            self.output_depth_dir = self.output_dir / "depth_maps"
+            self.output_depth_dir.mkdir(parents=True, exist_ok=True)
+
     def _set_temporary_directories(self):
         self.temp_dir = self.output_dir / "temp"
         self.temp_dir.mkdir(parents=True, exist_ok=True)
@@ -56,21 +70,19 @@ class DitchNetPredictor:
         self.hpmf_temp = self.temp_dir / "hpmf_temp.tif"
         self.isi_temp = self.temp_dir / "isi_temp.tif"
 
-    def _create_output_layer(self, feature_array):
-        # Define the expected input and output raster dimensions.
-        # The DEMs are processed as 2048x2048 grids, divided into 512x512 tiles.
-        original_raster_size = 2048
+    def _create_output_layer(self, feature_array, orig_height, orig_width, resampled_height, resampled_width):
         chip_size = 512
 
-        # Allocate memory for the full 2048x2048 output raster
-        output_array = np.empty((original_raster_size, original_raster_size), dtype=np.float32)
+        output_array = np.empty((resampled_height, resampled_width), dtype=np.float32)
 
         # Loop over the raster in steps of chip_size to create non-overlapping chips
-        for i in range(0, original_raster_size, chip_size):
-            for j in range(0, original_raster_size, chip_size):
+        for i in range(0, resampled_height, chip_size):
+            start_i = min(i, resampled_height - chip_size)
+            for j in range(0, resampled_width, chip_size):
+                start_j = min(j, resampled_width - chip_size)
 
                 # Extract a tile from the input features (2 channels: HPMF + ISI)
-                feature_chip = feature_array[:, i:i + chip_size, j:j + chip_size]
+                feature_chip = feature_array[:, start_i:start_i + chip_size, start_j:start_j + chip_size]
 
                 # Add batch dimension and convert to PyTorch tensor
                 feature_chip = feature_chip[np.newaxis, :, :]
@@ -84,39 +96,64 @@ class DitchNetPredictor:
                     predicted = torch.sigmoid(predicted).squeeze().cpu().numpy()
 
                 # Insert predicted tile back into the corresponding position
-                output_array[i:i + chip_size, j:j + chip_size] = predicted
+                output_array[start_i:start_i + chip_size, start_j:start_j + chip_size] = predicted
 
-        # Resize the combined output to the final 2000x2000 DEM resolution
-        output_array = resize(output_array, (2000, 2000), order=1, preserve_range=True, anti_aliasing=False)
+        output_array = resize(output_array, (orig_height, orig_width), order=1, preserve_range=True, anti_aliasing=False)
 
         return output_array
 
     def _output_probability_map(self, input_path, profile, output_array):
         profile.update(dtype=rasterio.float32, count=1, nodata=None)
-        with rasterio.open(self.output_probability_dir / f"{input_path.stem}_ditch_probability.tif", "w",
-                           **profile) as dst:
+
+        output_file = self.output_probability_dir / f"{input_path.stem}_ditch_probability.tif"
+        with rasterio.open(output_file, "w", **profile) as dst:
             dst.write(output_array, 1)
 
     def _output_classified_map(self, input_path, profile, output_array):
         filtered_output_array = (output_array >= self.threshold).astype(np.uint8)
 
         profile.update(dtype=rasterio.uint8, count=1, nodata=None)
-        with rasterio.open(self.output_classified_dir / f"{input_path.stem}_ditch_classified.tif", "w",
-                           **profile) as dst:
+
+        output_file = self.output_classified_dir / f"{input_path.stem}_ditch_classified.tif"
+        with rasterio.open(output_file, "w", **profile) as dst:
             dst.write(filtered_output_array, 1)
 
-    def _process_single_dem(self, dem_path):
-        # Read the DEM file metadata (e.g., CRS, transform, resolution)
-        with rasterio.open(dem_path) as src:
+    def _output_depth_map(self, input_path, output_array):
+        with rasterio.open(self.hpmf_temp) as src:
+            hpmf_ref = src.read(1)
             profile = src.profile
+
+        hpmf_ref[hpmf_ref == -9999] = 0
+        output_array = np.where(output_array > 0.1, hpmf_ref, 0)
+        output_array[output_array > 0] = 0
+
+        output_file = self.output_depth_dir / f"{input_path.stem}_ditch_depth.tif"
+        with rasterio.open(output_file, "w", **profile) as dst:
+            dst.write(output_array, 1)
+
+    def _process_single_dem(self, dem_path):
+        # Read the DEM file metadata
+        with rasterio.open(dem_path) as src:
+            orig_height = src.height
+            orig_width = src.width
+            profile = src.profile
+
+        if orig_height < 512 or orig_width < 512:
+            print(f"Input image {dem_path.name} is too small — minimum size is 512x512 pixels.")
+            self.invalid_inputs.append(dem_path.name)
+            return
 
         # Generate High-Pass Median Filter and Impoundment Size Index layers
         hpmf_array = create_hpmf_layer(dem_path, self.hpmf_temp)
         isi_array = create_isi_layer(dem_path, self.isi_temp)
 
+        resampled_height = int(orig_height + (orig_height * 0.024))
+        resampled_width = int(orig_width + (orig_width * 0.024))
+
         # Combine layers into a 2-channel feature array for the model
-        feature_array = create_feature_layer(hpmf_array, isi_array)
-        output_array = self._create_output_layer(feature_array)
+        feature_array = create_feature_layer(hpmf_array, isi_array, resampled_height, resampled_width)
+        output_array = self._create_output_layer(feature_array, orig_height, orig_width,
+                                                 resampled_height, resampled_width)
 
         # Save probability and/or classified maps depending on user options
         if self.create_prob_map:
@@ -125,38 +162,82 @@ class DitchNetPredictor:
         if self.create_class_map:
             self._output_classified_map(dem_path, profile, output_array)
 
+        if self.create_depth_map:
+            self._output_depth_map(dem_path, output_array)
+
         # Clean up temporary files created by WhiteboxTools
-        self.hpmf_temp.unlink()
-        self.isi_temp.unlink()
+        for temp in (self.hpmf_temp, self.isi_temp):
+            if temp.exists():
+                temp.unlink()
+
+    def _create_virtual_rasters(self):
+        if self.create_prob_map:
+            depth_raster_files = [raster_name for raster_name in self.output_probability_dir.glob("*.tif")]
+
+            vrt_path = self.output_probability_dir / "ditch_probability_map.vrt"
+            gdal.BuildVRT(vrt_path, depth_raster_files)
+
+        if self.create_class_map:
+            depth_raster_files = [raster_name for raster_name in self.output_classified_dir.glob("*.tif")]
+
+            vrt_path = self.output_classified_dir / "ditch_classified_map.vrt"
+            gdal.BuildVRT(vrt_path, depth_raster_files)
+
+        if self.create_depth_map:
+            depth_raster_files = [raster_name for raster_name in self.output_depth_dir.glob("*.tif")]
+
+            vrt_path = self.output_depth_dir / "ditch_hpmf_depth_map.vrt"
+            gdal.BuildVRT(vrt_path, depth_raster_files)
 
     def predict(self):
-        print(f"Running DitchNet inference on DEM files in: {self.input_dem_dir}")
+        print(f"Running DitchNet inference on DEM files in: {self.input_dem_dir}\n")
 
         dem_files = list(self.input_dem_dir.glob("*.tif"))
         if not dem_files:
             print("No DEM (.tif) files found — nothing to process.")
-            shutil.rmtree(self.temp_dir)
+
+            if self.output_probability_dir.exists():
+                shutil.rmtree(self.output_probability_dir)
+
+            if self.output_classified_dir.exists():
+                shutil.rmtree(self.output_classified_dir)
+
+            if self.output_depth_dir.exists():
+                shutil.rmtree(self.output_depth_dir)
+
+            if self.temp_dir.exists():
+                shutil.rmtree(self.temp_dir)
+
             return
 
         for dem_path in dem_files:
             print(f"Processing: {dem_path.name}")
             self._process_single_dem(dem_path)
 
-        shutil.rmtree(self.temp_dir)
+        self._create_virtual_rasters()
+
+        if self.temp_dir.exists():
+            shutil.rmtree(self.temp_dir)
+
         print(f"All predictions completed.")
+
+        if self.invalid_inputs:
+            print(f"\nThe following input files failed to process:")
+            for dem_name in self.invalid_inputs:
+                print(dem_name)
 
 
 class Main:
     def __init__(self):
         self.args = self._parse_arguments()
-
         model = DitchNet.load_from_checkpoint(self.args.model_path)
         self.predictor = DitchNetPredictor(model,
                                            self.args.input_dem_dir,
                                            self.args.output_dir,
                                            self.args.threshold,
                                            self.args.output_prob_map,
-                                           self.args.output_class_map)
+                                           self.args.output_class_map,
+                                           self.args.output_depth_map)
         self.run()
 
     @staticmethod
@@ -182,6 +263,11 @@ class Main:
                             dest="output_class_map",
                             action="store_false",
                             help="Disable saving of the classified map output (enabled by default).")
+
+        parser.add_argument("--no_depth_map",
+                            dest="output_depth_map",
+                            action="store_false",
+                            help="Disable saving of the depth map output (enabled by default).")
 
         return parser.parse_args()
